requirements.txt
openai
langchain-nvidia-ai-endpoints
langchain_community
faiss-cpu
python-dotenv
streamlit
pypdf
